+++ CHANGELOG.md
@@ -1,5 +1,11 @@
-Laravel Composite Validation Repository
-=======================================
+Laravel Composite Validation
+============================
+
+1.0.1 under development
+-----------------------
+
+- Bug #3: Fixed `CompositeRule` skipped on array attributes like 'item_ids.*' or 'items.*.id' (klimov-paul)
+
 
 1.0.0, June 18, 2019
 --------------------